requirements.txt
dash==2.7.0
pandas==1.5.2
pydantic==1.10.2
redis==4.3.5
Flask_SQLAlchemy==3.0.2
dash-bootstrap-components==1.2.1
flask_caching==2.0.1
flask_session==0.4.0
Flask_Login==0.6.2
pyscss==1.4.0
csscompressor==0.9.5
wheezy.routing==3.0.1
faker==15.3.4
sqlalchemy_serializer==1.4.1
requests==2.28.1
polars==0.15.2
email_validator==1.3.0
python_dotenv==0.21.0